dom/media/test/test_bug895091.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Bug 895091 - Integrating vtt.js</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<video id="v" src="seek.webm" preload="metadata">
  <track src="long.vtt" kind="subtitles" id="track1">
  <track src="long.vtt" kind="subtitles" id="track2">
</video>
<script type="text/javascript">
SimpleTest.waitForExplicitFinish();
let video = document.getElementById("v");
video.addEventListener("loadedmetadata", function run_tests() {
  let trackElement = document.getElementById("track1");
  let trackElementTwo = document.getElementById("track2");
  // Re-que run_tests() at the end of the event loop until the track
  // element has loaded its data.
  if (trackElement.readyState == 1 || trackElementTwo.readyState == 1) {
    setTimeout(run_tests, 0);
    return;
  }

  // Set mode to hidden so that the active cue lists are being updated.
  trackElement.track.mode = "hidden";
  trackElementTwo.track.mode = "hidden";

  is(trackElement.readyState, 2, "Track::ReadyState should be set to LOADED.");
  is(trackElementTwo.readyState, 2, "Track::ReadyState should be set to LOADED.");

  is(trackElement.track.cues.length, 2000, "Cue list length should be 2000.");
  is(trackElementTwo.track.cues.length, 2000, "Cue list length should be 2000.");

  SimpleTest.finish();
});
</script>
</body>
</html>
